Update player_template.py
I tried to put a spin on the player template and added some voice lines and tried to put in a shop menu and a character statistics. I think I tried to do more than I can handle so far... The code runs but, it is stuck in a loop and I have no idea how to break the loop.